chore: split EditableText
