+++ packages/components/src/EditableText/PlainTextTitle.module.scss
@@ -1,0 +1,47 @@
+@use '~@talend/bootstrap-theme/src/theme/guidelines' as *;
+
+@mixin ellipsis {
+	white-space: nowrap;
+	text-overflow: ellipsis;
+	overflow: hidden;
+}
+
+.tc-editable-text-title {
+	display: flex;
+	align-items: center;
+	max-width: 90rem;
+
+	:global(.tc-editable-text-wording-wrapper) {
+		@include ellipsis;
+	}
+
+	.tc-editable-text-wording {
+		white-space: nowrap;
+		text-overflow: ellipsis;
+		overflow: hidden;
+	}
+	.tc-editable-text-pencil {
+		position: relative;
+		left: -2000px;
+		padding-left: $padding-smaller;
+		padding-right: $padding-smaller;
+		flex: 0 0 auto;
+		line-height: inherit;
+		min-height: inherit;
+		color: $white;
+
+		&:global(.btn-link) {
+			color: $gray;
+		}
+
+		&:focus,
+		&:hover {
+			position: static;
+		}
+	}
+	&:hover {
+		.tc-editable-text-pencil {
+			position: static;
+		}
+	}
+}
